+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,67 +1,13 @@
 package com.techcourse.service;
 
-import com.interface21.dao.DataAccessException;
-import com.interface21.jdbc.datasource.DataSourceUtils;
-import com.interface21.transaction.support.TransactionSynchronizationManager;
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
-import java.sql.Connection;
-import java.sql.SQLException;
-import javax.sql.DataSource;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-public class UserService {
+public interface UserService {
 
-    private static final Logger log = LoggerFactory.getLogger(UserService.class);
+    User findById(final long id);
 
-    private final DataSource dataSource;
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
+    void save(final User user);
 
-    public UserService(final DataSource dataSource, final UserDao userDao, final UserHistoryDao userHistoryDao) {
-        this.dataSource = dataSource;
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-    }
-
-    public User findById(final long id) {
-        return userDao.findById(id);
-    }
-
-    public void insert(final User user) {
-        userDao.insert(user);
-    }
-
-    public void changePassword(final long id, final String newPassword, final String createBy) {
-        Connection connection = null;
-        try {
-            connection = DataSourceUtils.getConnection(dataSource);
-            connection.setAutoCommit(false);
-
-            User user = findById(id);
-            user.changePassword(newPassword);
-
-            userDao.update(user);
-            userHistoryDao.log(new UserHistory(user, createBy));
-
-            connection.commit();
-        } catch (Exception e) {
-            if (connection != null) {
-                try {
-                    connection.rollback();
-                } catch (SQLException rollbackEx) {
-                    log.error("Rollback failed", rollbackEx);
-                }
-            }
-            throw new DataAccessException(e);
-        } finally {
-            if (connection != null) {
-                DataSourceUtils.releaseConnection(connection, dataSource);
-                TransactionSynchronizationManager.unbindResource(dataSource);
-            }
-        }
-    }
+    void changePassword(final long id, final String newPassword, final String createdBy);
 }
+
